Moved general.kmean in performance
