+++ openquake/baselib/tests/general_test.py
@@ -19,14 +19,13 @@
 """
 Test related to code in openquake/utils/general.py
 """
-import time
 import unittest.mock as mock
 import unittest
 import numpy
 from operator import attrgetter
 from collections import namedtuple
 from openquake.baselib.general import (
-    block_splitter, split_in_blocks, assert_close, kmean, rmsdiff,
+    block_splitter, split_in_blocks, assert_close, rmsdiff,
     deprecated, DeprecationWarning, cached_property, compress, decompress)
 
 
@@ -204,44 +203,6 @@ class CompressTestCase(unittest.TestCase):
     def test(self):
         a = dict(a=numpy.array([9999.]))
         self.assertEqual(a, decompress(compress(a)))
-
-
-class KmeanTestCase(unittest.TestCase):
-    def test_small(self):
-        # build a small structured array
-        dtlist = [('mdbin', numpy.uint32), ('rake', numpy.float64),
-                  ('sids', numpy.uint32)]
-        N = 10
-        arr = numpy.zeros(N, dtlist)
-        rng = numpy.random.default_rng(42)
-        arr['mdbin'] = rng.integers(50, size=N)
-        arr['rake'] = rng.random(N) * 360
-        arr['sids'] = rng.integers(1000, size=N)
-        sids = []
-        for rec in kmean(arr, 'mdbin'):
-            sids.append(arr['sids'][arr['mdbin'] == rec['mdbin']])
-        expected_sids = [[450, 858, 631], [276], [554, 887], [92],
-                         [827], [227], [63]]
-        numpy.testing.assert_equal(sids, expected_sids)
-
-    def test_big(self):
-        # build a very large structured array
-        dtlist = [('mdbin', numpy.uint32), ('rake', numpy.float64),
-                  ('sids', numpy.uint32)]
-        N = 10_000_000
-        arr = numpy.zeros(N, dtlist)
-        rng = numpy.random.default_rng(42)
-        arr['mdbin'] = rng.integers(50, size=N)
-        arr['rake'] = rng.random(N) * 360
-        arr['sids'] = rng.integers(1000, size=N)
-        t0 = time.time()
-        mean = kmean(arr, 'mdbin')
-        sids = []
-        for mdbin in mean['mdbin']:
-            sids.append(arr['sids'][arr['mdbin'] == mdbin])
-        print([len(s) for s in sids])
-        dt = time.time() - t0
-        print('Grouped %d elements in %.1f seconds' % (N, dt))
 
 
 class RmsDiffTestCase(unittest.TestCase):
