+++ openquake/baselib/tests/performance_test.py
@@ -19,7 +19,7 @@ import time
 import unittest
 import pickle
 import numpy
-from openquake.baselib.performance import Monitor, split_array
+from openquake.baselib.performance import Monitor, split_array, kmean
 
 
 class MonitorTestCase(unittest.TestCase):
@@ -79,3 +79,41 @@ class SplitArrayTestCase(unittest.TestCase):
         expected_sids = [[631, 858, 450], [276], [887, 554], [92],
                          [827], [227], [63]]
         numpy.testing.assert_equal(sids, expected_sids)
+
+
+class KmeanTestCase(unittest.TestCase):
+    def test_small(self):
+        # build a small structured array
+        dtlist = [('mdbin', numpy.uint32), ('rake', numpy.float64),
+                  ('sids', numpy.uint32)]
+        N = 10
+        arr = numpy.zeros(N, dtlist)
+        rng = numpy.random.default_rng(42)
+        arr['mdbin'] = rng.integers(50, size=N)
+        arr['rake'] = rng.random(N) * 360
+        arr['sids'] = rng.integers(1000, size=N)
+        sids = []
+        for rec in kmean(arr, 'mdbin'):
+            sids.append(arr['sids'][arr['mdbin'] == rec['mdbin']])
+        expected_sids = [[450, 858, 631], [276], [554, 887], [92],
+                         [827], [227], [63]]
+        numpy.testing.assert_equal(sids, expected_sids)
+
+    def test_big(self):
+        # build a very large structured array
+        dtlist = [('mdbin', numpy.uint32), ('rake', numpy.float64),
+                  ('sids', numpy.uint32)]
+        N = 10_000_000
+        arr = numpy.zeros(N, dtlist)
+        rng = numpy.random.default_rng(42)
+        arr['mdbin'] = rng.integers(50, size=N)
+        arr['rake'] = rng.random(N) * 360
+        arr['sids'] = rng.integers(1000, size=N)
+        t0 = time.time()
+        mean = kmean(arr, 'mdbin')
+        sids = []
+        for mdbin in mean['mdbin']:
+            sids.append(arr['sids'][arr['mdbin'] == mdbin])
+        print([len(s) for s in sids])
+        dt = time.time() - t0
+        print('Grouped %d elements in %.1f seconds' % (N, dt))
